Feature/integrated calendar

+++ integrated_calendar/static/script/interface/calendar-interface.js
@@ -407,8 +407,26 @@ async function taskPreviewInterface(d) {
 
                 delete_btn.addEventListener("mouseup", async () => {
                     const reminder = new Reminder(null, c.id);
-                    await reminder.delete();
-                    tasks_holder.removeChild(div);
+
+                    await Swal.fire({
+                        title: 'Are you sure you want to delete?',
+                        text: "You won't be able to revert this!",
+                        icon: 'warning',
+                        showCancelButton: true,
+                        confirmButtonColor: '#3085d6',
+                        cancelButtonColor: '#d33',
+                        confirmButtonText: 'Yes, delete it!'
+                    }).then(async (result) => {
+                        if (result.isConfirmed) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+                            await reminder.delete();
+                            tasks_holder.removeChild(div);
+                        }
+                    });
                 });
 
                 update_btn.addEventListener("mouseup", async () => {
